Fixed error on preview image
Fixed error on non-unlimited mode

+++ app/src/main/java/com/thibaudperso/sonycamera/timelapse/ui/processing/ProcessingFragment.java
@@ -453,7 +453,7 @@ public class ProcessingFragment extends Fragment {
                     double ratio = (double) totalElapsedTime / totalTime;
                     mOverallProgressBar.setProgress((int) totalElapsedTime);
                     mOverallProgressValue.setText(String.format(getString(R.string.percent1f),
-                            (int) (ratio * 100)));
+                            ratio * 100));
                 }
             }
